Add listbox and edit functionality to GUI
The GUI was updated to include a listbox which displays current todos. The script now fetches the existing todos at startup and lists them in the new listbox. Also, an "Edit" button was added to allow the user to select a todo from the list and edit its content. The edited todo is then updated in the list and saved to the disk.